Switched to MVC
works
uses InMemoryRepository due to Hibernate one doesn't work properly
No frontend yet
