+++ src/main/java/org/aleksid/wikime/Application.java
@@ -1,0 +1,13 @@
+package org.aleksid.wikime;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Application {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+
+}
