+++ src/main/resources/application.properties
@@ -1,0 +1,11 @@
+spring.jpa.hibernate.ddl-auto=update
+spring.datasource.url=jdbc:postgresql://localhost:5432/wikime_hiber
+spring.datasource.username=wikime
+spring.datasource.password=wikime
+#spring.jpa.generate-ddl=true
+
+#spring.freemarker.expose-request-attributes=true
+
+upload.path = /C:/Users/Alexandr/IdeaProjects/wikime/uploads
+
+security.ignored=all
