+++ src/main/resources/templates/parts/common.ftlh
@@ -1,0 +1,12 @@
+<#macro page>
+    <!DOCTYPE html>
+    <html lang="en">
+    <head>
+        <meta charset="UTF-8">
+        <title>Wikime</title>
+    </head>
+    <body>
+    <#nested>
+    </body>
+    </html>
+</#macro>
