+++ src/main/resources/templates/parts/util.ftlh
@@ -1,0 +1,19 @@
+<#macro  util>
+    <div>
+        <table>
+            <tr>
+                    <td><button onclick="window.history.back()">Back</button></td>
+                    <td>
+                        <form action="/wikime" method="get">
+                            <button type="submit">Home</button>
+                        </form>
+                    </td>
+                    <td>
+                        <form action="/articles" method="get">
+                            <button type="submit">Main</button>
+                        </form>
+                    </td>
+            </tr>
+        </table>
+    </div>
+</#macro>
