+++ src/test/java/HibernateTest2dir.java
@@ -1,6 +1,5 @@
-import com.sun.xml.bind.v2.TODO;
-import dto.DBArticle;
-import dto.DBTag;
+import org.aleksid.wikime.dto.DBArticle;
+import org.aleksid.wikime.dto.DBTag;
 import org.hibernate.Criteria;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
